# aria-frontend/openapi/aria.json
{
  "components": {
    "schemas": {
...
      "BrandProfileRecord": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/BrandProfile"
          },
          "profile_version": {
            "title": "Profile Version",
            "type": "integer"
          },
          "updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "validation": {
            "$ref": "#/components/schemas/BrandProfileValidationResult"
          }
        },
        "required": [
          "profile",
          "validation",
          "profile_version"
        ],
        "title": "BrandProfileRecord",
        "type": "object"
      },
...
      "BrandProfileUpdate": {
        "properties": {
          "expected_version": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Version"
          },
          "profile": {
            "$ref": "#/components/schemas/BrandProfile"
          }
        },
        "required": [
          "profile"
        ],
        "title": "BrandProfileUpdate",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "ARIA LLM Orchestration Service",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/audit": {
      "get": {
        "operationId": "audit_v1_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "type": "object"
                  },
                  "title": "Response Audit V1 Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audit",
        "tags": [
          "product"
        ]
      }
    },
    "/v1/brands/{brand_id}/profile": {
      "get": {
        "operationId": "get_brand_profile_v1_brands__brand_id__profile_get",
        "parameters": [
          {
            "in": "path",
            "name": "brand_id",
            "required": true,
            "schema": {
              "title": "Brand Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandProfileRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Brand Profile",
        "tags": [
          "product"
        ]
      },
      "put": {
        "operationId": "put_brand_profile_v1_brands__brand_id__profile_put",
        "parameters": [
          {
            "in": "path",
            "name": "brand_id",
            "required": true,
            "schema": {
              "title": "Brand Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandProfileRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Put Brand Profile",
        "tags": [
          "product"
        ]
      }
    },
    "/v1/brands/{brand_id}/profile/validate": {
      "post": {
        "operationId": "validate_brand_profile_v1_brands__brand_id__profile_validate_post",
        "parameters": [
          {
            "in": "path",
            "name": "brand_id",
            "required": true,
            "schema": {
              "title": "Brand Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BrandProfileUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrandProfileRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Validate Brand Profile",
        "tags": [
          "product"
        ]
      }
    },
    "/v1/calendar/items": {
      "get": {
        "operationId": "list_calendar_v1_calendar_items_get",
        "parameters": [
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Platform"
            }
          },
          {
            "in": "query",
            "name": "planning_state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Planning State"
            }
          },
          {
            "in": "query",
            "name": "approval_status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Approval Status"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalendarPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Calendar",
        "tags": [
          "product"
        ]
      },
      "post": {
        "operationId": "create_calendar_item_v1_calendar_items_post",
        "parameters": [
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CalendarCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Create Calendar Item V1 Calendar Items Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Calendar Item",
        "tags": [
          "product"
        ]
      }
    },
...
    "/v1/capabilities": {
      "get": {
        "operationId": "capabilities_v1_capabilities_get",
        "parameters": [
          {
            "in": "header",
            "name": "X-ARIA-Workspace-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Workspace-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Role",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Role"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-User-ID",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-User-Id"
            }
          },
          {
            "in": "header",
            "name": "X-ARIA-Test-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Aria-Test-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Capabilities",
        "tags": [
          "product"
        ]
      }
    },
...
  }
}